Partial io ports overhaul + more sample clipping improvements.

+++ macosx/PortableAudioClip.h
@@ -108,13 +108,13 @@ static inline SInt32 FloatToInt32(const double val){
 }
 
 static inline double fabs(const double val){
- 
+	
 #if defined(PPC) && !defined(NO_ASM)
     double d;
     
     __asm__ volatile ("fabs %0, %1"
-             : "=f" (d)
-             : "f" (val));
+					  : "=f" (d)
+					  : "f" (val));
     
     return d;
 #else
@@ -137,13 +137,17 @@ static inline double clamp(const double val){
 
 
 static void Float32ToSInt32_portable( const float* floatMixBuf, SInt32* destBuf, const size_t end, const size_t start){
-    register float inSample;
+    register double inSample;
     register size_t sampleIndex = start, sampleDestinationMemoryIndex = (start * sizeof(*destBuf));
     
     // Loop through the mix/sample buffers one sample at a time and perform the clip and conversion operations
     for (; sampleIndex < end; sampleIndex++, sampleDestinationMemoryIndex += sizeof(*destBuf)){
         // Fetch the floating point mix sample
-        inSample = (clamp((const double)floatMixBuf[sampleIndex]) * clipPosMul32) + 128;
+        inSample = (
+						clamp(
+							floatMixBuf[sampleIndex]
+						) * clipPosMul32
+					) + 128;
         
         // Scale the -1.0 to 1.0 range to the appropriate scale for signed 32-bit samples and then
         // convert to SInt32 and store in the hardware sample buffer
@@ -154,13 +158,13 @@ static void Float32ToSInt32_portable( const float* floatMixBuf, SInt32* destBuf,
 }
 
 static void Float32ToSInt16Aligned32_portable( const float* floatMixBuf, SInt32* destBuf, const size_t end, const size_t start){
-    register float inSample;
+    register double inSample;
     register size_t sampleIndex = start, sampleDestinationMemoryIndex = (start * sizeof(*destBuf));
     
     // Loop through the mix/sample buffers one sample at a time and perform the clip and conversion operations
     for (; sampleIndex < end; sampleIndex++, sampleDestinationMemoryIndex += sizeof(*destBuf)){
         // Fetch the floating point mix sample
-        inSample = (clamp((const double)floatMixBuf[sampleIndex]) * clipPosMul16);
+        inSample = (clamp(floatMixBuf[sampleIndex]) * clipPosMul16);
         
         // Scale the -1.0 to 1.0 range to the appropriate scale for signed 16-bit samples and then
         // convert to SInt16 and shit upwards of 16 bits (to have the correct volume) and store in the hardware sample buffer
@@ -171,13 +175,13 @@ static void Float32ToSInt16Aligned32_portable( const float* floatMixBuf, SInt32*
 }
 
 static void Float32ToSInt16_portable( const float* floatMixBuf, SInt16* destBuf, const size_t end, const size_t start){
-    register float inSample;
+    register double inSample;
     register size_t sampleIndex = start, sampleDestinationMemoryIndex = (start * sizeof(*destBuf));
     
     // Loop through the mix/sample buffers one sample at a time and perform the clip and conversion operations
     for (; sampleIndex < end; sampleIndex++, sampleDestinationMemoryIndex += sizeof(*destBuf)){
         // Fetch the floating point mix sample
-        inSample = (clamp((const double)floatMixBuf[sampleIndex]) * clipPosMul16);
+        inSample = (clamp(floatMixBuf[sampleIndex]) * clipPosMul16);
         
         // Scale the -1.0 to 1.0 range to the appropriate scale for signed 16-bit samples and then
         // convert to SInt16 and store in the hardware sample buffer
@@ -287,6 +291,5 @@ static void Float32ToSInt16_no_array_portable( const float* mixBuf, SInt16* dest
 #define Float32ToSInt16Aligned32_optimized Float32ToSInt16Aligned32_portable
 
 #endif
-
 
 #endif
